+++ shape/design/classdia_test.go
@@ -1,7 +1,16 @@
 package design
 
-import "testing"
+import (
+	"testing"
 
-func TestClassDiagram_OtherType(t *testing.T) {
+	"github.com/gregoryv/draw/shape"
+)
 
+func TestClassDiagram(t *testing.T) {
+	t.Run("InlineSVG", func(t *testing.T) {
+		d := NewClassDiagram()
+		record := d.Struct(shape.Record{})
+		d.Place(record).At(20, 20)
+		checkInlining(t, d)
+	})
 }