+++ shape/design/package.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"os"
 
+	"github.com/gregoryv/draw"
 	"github.com/gregoryv/draw/shape"
 )
 
@@ -21,4 +22,9 @@ func saveAs(dia svgWriter, style shape.Style, filename string) error {
 	defer fh.Close()
 	style.SetOutput(fh)
 	return dia.WriteSVG(&style)
+}
+
+func inlineSVG(w io.Writer, d draw.SVGWriter, s *shape.Style) error {
+	s.SetOutput(w)
+	return d.WriteSVG(s)
 }