+++ shape/design/package_test.go
@@ -1,8 +1,10 @@
 package design
 
 import (
+	"bytes"
 	"testing"
 
+	"github.com/gregoryv/draw"
 	"github.com/gregoryv/draw/shape"
 )
 
@@ -11,4 +13,17 @@ func Test_saveAs(t *testing.T) {
 	if err == nil {
 		t.Fail()
 	}
+}
+
+func checkInlining(t *testing.T, d draw.SVGInlineWriter) (inlined, plain bytes.Buffer) {
+	t.Helper()
+	d.InlineSVG(&inlined)
+	d.WriteSVG(&plain)
+
+	if inlined.String() == plain.String() {
+		t.Error("Inlined same as plain")
+		t.Log("inlined: \n", inlined.String())
+		t.Log("plain: \n", plain.String())
+	}
+	return
 }